[FIX] worksheet, industry_fsm_report: worksheet report's arch handles only boolean invisible
On a worksheet template, click on Design Template. You arrive in studio, editing the form view for the template
Add some fields before the already existing comments field.
Remove the comment field.
Rename the last added field.
Move that field on top of the other added fields.

Before this commit, the moved field was never at the desired spot.
This is because of the `def normalize` in web_studio. When it tries to determine simple xpath
for the operations, it gets lost when a field is replaced by other stuff. Specifically, it can't really handle
emptying a node from existing fields and filling that with new things.

This is a limitation that could be fixed at the cost of heavy work and potential bugs created by an hypothetical fix.
Considering the sensitivity of the `normalize` function, This commit proposes to handle things differently in worksheet.

It now supports having an invisible field in the form view, that will have the consequence of not putting the field in the
QWeb report for that template model.

For simplicity's sake, it only supports Boolean values.

opw-3601040

closes odoo/enterprise#54303

X-original-commit: 2f37797cd08422967957bbb8bb78fbd3c488b0b2
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>