[IMP] whatsapp: keep same channel for same customer / user
PURPOSE

Currently if a channel is active on WhatsApp and the same user sends a
template on another model then on further conversation a new channel
is created. For the same user it would be better to keep the same
channel.

SPECIFICATIONS

If user sends a template and Customer / subscriber replies to that
a channel is created and conversation is active. After that a new
template is send by the current user, post this if the customer replies
it should not create a new channel and the conversation should continue
in the same channel.

task-3644180

closes odoo/enterprise#54327

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>