[IMP] web_studio: enable 'string' and 'class' attribute edition
This commit brings new attributes to the sidebar when editing a button node.
Before this commit, it was not possible to change the label of a button
without editing the XML, and it was not possible to change the color class
being used on the button.

Now, an input is present to change the string attribute directly from the UI.
Another inout has been added below, to choose and edit from any bootstrap class.

A test has been added to verify those attributes are now editable from the user
interface.

task-3599683

closes odoo/enterprise#54285

Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>