[FIX] l10n_mx_edi: add ondelete cascade to certificate company_id field
Currently, you cannot delete a certificate.

Steps to reproduce
------------------
- install `l10n_mx_edi`
* switch to a Mexican company
* go to the settings
* under 'MX Electronic invoicing', delete a line from the 'Certificates'
  field
* attempt to save

You should be met with a validation error.

Cause
-----
In the `Certificate` model, the `company_id` field is defined as a
`Many2one` relation to the `ResCompany` model and is marked as
`required=True`. The inverse relation in the `ResCompany` model contains
a `certificate_ids` field.

When saving after removing a certificate in the settings, Odoo attempts
to unlink the certificate from `company_id.certificate_ids`. Due to the
defined relationships, this will result in the system trying to set the
`certificate.company_id` field to `False`. However, since `company_id`
is a required field, this operation leads to an error.

opw-3664517

closes odoo/enterprise#54134

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>