[FIX] web_studio: display error when edition fails
CONTEXT
=======
On a ViewEditor, make a change that will result in an error server side.
There are two main reasons for that to happen:
- an xpath was faulty and ir.ui.view crashes in `def apply_inheritance_specs`
- the view check raises because a field is not present, or some other node depends on another and is not satisfied
the crash could be in `ir.ui.view:def _check_xml` or `ir.ui.view:def check`

Unfortunately, all cases use the def _raise_view_function that raises an error with some metadata
that are useless for us to provide the user with a an elegant message that could describe
a way around the error, if possible.

FIX
===
Before this commit any error raised by def edit_view was hidden in the client, only a notification appeared on the screen.

This was confusing because the user had no clue as to what to do next.

After this commit, we re-throw the error in client after having reverted the changes in the ViewEditor

Hopefully in the future, we would have better error messages or even an autocorrect heuristics.

closes odoo/enterprise#54363

X-original-commit: a0a688b98a625e04752b9f9dd6d98d36cfd14b7a
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>