[FIX] account_intrastat: fix report for product coming from UK
Steps to reproduce:
- Install account_intrastat
- Install a localization that should declare intrastat report (e.g. l10n_be)
- Switch to a Belgian company
- Create a product with the country of origin set to United Kingdom
- Create an invoice with that product and set Intrastat country to
any country except United Kingdom
- Confirm the invoice
- Go to "Accounting / Reporting / Audit Reports / Intrastat Report"
- Filter on the period of the invoice
- Export in XML

Issue:
The country code for United Kingdom in the report is GB.
However, since Brexit, 2 different codes should be used for United Kingdom (XU)
and Northen Ireland (XI).
Since 2024, GB is not accepted anymore by the government and the intrastat report
is rejected if XU/XI are not used instead.

Solution:
This fix adds Northern Ireland as a country, to be able to select it as a country
of origin of a product.
In the intrastat report, if the country code is GB, it will be changed to XU.

opw-3669871

closes odoo/enterprise#54618

X-original-commit: bfa2da11d683f774ddb975157600e25341b96090
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>