[FIX] spreadsheet_dashboard_sale_subscription: add report form view
Let's say a `sale.subscription.report` list is inserted into a spreadsheet
(e.g. in the Subscriptions dashboard). Right click on one of the list
cells and click on "See record". It opens an auto-generated form view which
is unusable.

Ideally we would want to be redirected to the related sale order form view.
That's what is done when click on a `sale.subscription.report` in the
list view. It's done with `action="action_open_subscription_order"` on
the list arch. Unfortunatly, we don't have such a mechanism to overwrite
the action for lists in spreadsheet.

So, in this commit, we introduce a minimal form view allowing to go
the sale order, without having to find the field among the many many fields
in the auto-generated form view.

closes odoo/enterprise#54408

Task: 3680601
X-original-commit: b5a157f5843b58bc6560bdbc6f6746f5f766bdb0
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>