[FIX] whatsapp: fix retrying message with report on template
This traceback occurs when the user retries a failed message that has a
template with a report defined on it.

Steps to reproduce

- Create a template with a header-type document and set a report in it
- Try to send a message with the template created and it must fail
- Make sure the failed message must be retryable. (e.g. try to fail message
  using `132001` error code)
- Try to resend this message
- Error will be generated in the backend

Problem

`_get_send_template_vals` is responsible for returning `send_vals` and
`attachment` which will be sent in the message. If the attachment is sent in the
parameters of `_get_send_template_vals` then will return the same attachment
that was sent in the parameters. And if no attachment is sent then it will
return the header attachment or it will return the generated attachment from the
report. In the current scenario, while retrying the failed message whose report
was previously generated, `_get_send_template_vals` is called first, and
`whatsapp_message.mail_message_id.attachment_ids` is sent in a parameter that is
an old attachment, so it will return the old attachment which will be assigned
to `attachment`. After this, `whatsapp_message.mail_message_id.attachment_ids`
is unlinked, and then the `attachment` which is still holding the old attachment
record is linked. As the old attachment is unlinked while linking it again it
will give an error of attachment is not present in table `ir_attachment`.

Solution:

Now, If the message having a report is retried then firstly
`whatsapp_message.mail_message_id.attachment_ids` which is an old attachment is
unlinked, and then `_get_send_template_vals` is called. As no attachment is sent
in the parameters of `_get_send_template_vals` it will return a newly generated
report in the form of a new attachment. The new attachment returned is linked
after this.

Task-3644990 (Whatsapp: Fix retrying message with report on template)

X-original-commit: odoo/enteprrise@e58c5e5506c5a644d28c36ada3c883f1e7dd2a27
Part-of: odoo/enterprise#59987