[FIX] spreadsheet_dashboard_sale_subscription: align charts
In the "MRR Evolution" dasbhoard, the data in the two
"MRR over time" charts are not aligned because of the
legend on the right on the second chart.
By "aligned", I mean September 2023 on the second chart
is just below September 2023 on the first chart.

It's easier to relate the two data for the two charts
when it's aligned.

Because I moved the legend on the bottom, I increased the
total height of the chart to leave enough space on the
vertical axis for the actual chart.

closes odoo/enterprise#54236

Task: 3680601
X-original-commit: 355b2f9bbf0790d6325b8ecc895950e66b7e9683
Related: odoo/odoo#149274
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>