[IMP] spreadsheet_edition: enforce t-inherit=t-name
The PR https://github.com/odoo/odoo/pull/145602 brings several changes
in the way inheritance of static templates is applied. One of which is
that for a template which inherits from a template with
t-name=template_name, the value of t-inherit should be exactly
template_name. This means we no longer accept implicit references to
modules names (in any ways). Here we adapt the code of
spreadsheet_edition to that particular change.

Task Id: 3432359

closes odoo/enterprise#53794

Related: odoo/odoo#145602
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>