[FIX] web_studio: fix related monetary field creation
Before this commit, creating a related field toward a monetary field
caused a crash. Since #40597, Studio handles monetary field differently
and related field has not been taken into account.

With this commit, it is now possible to create a related monetary field.
This commit also removes the currency dialog which is not longer used
thanks to this fix.

Steps to reproduce:
- Open Sales and a record
- Open studio
- Drag&Drop a new related field
- Select Customer->Total received field and confirm
- Crash

opw-3663307

closes odoo/enterprise#54803

X-original-commit: 967ac05cf6e5ba61bf3e21ee8cf286616cc26508
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>