[FIX] account_consolidation: properly audit consolidation report amounts with multiple journal lines
When an amount in the consolidation report was computed from multiple consolidation.journal.line objects, auditing it didn't work well: only the move lines corresponding to the first of those journal lines were shown.

This is due to https://github.com/odoo/odoo/blame/16.0/addons/web/static/src/search/search_arch_parser.js#L134 . Passing a default value for a filter consisting in a list of ids worked by chance and only in certain contexts in older versions. In 16.0, it does not anymore. This case is currently unsupported by the web framework (it will be on future versions; a task has been created for master) ; the alternative is here to directly rely on a domain on the action.

OPW 3662905

closes odoo/enterprise#54279

X-original-commit: efa8598b32a9425fc0cff639f0ceb8369c499b60
Signed-off-by: John Laterre (jol) <jol@odoo.com>