[FIX] appointment: fix access error on fetch
Purpose
=======
Fix the "_fetch_and_check_private_appointment_types" method
which was raising an Access Error when the user didn't have
access to at least one of the recordset record.

Specifications
==============
As the "appointment_type_ids" are fetch using a sudo, the ACLs
arent't applied to the fetched recordset which results in an
Access Error due to the user access rules.
Instead of checking the records access rules, filtering the recordset
will ensure to the user an access to the available ones.

Task-3544915

closes odoo/enterprise#49719

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>