[FIX] data_merge: manage the '%' char in the search company
Steps to reproduce:
-------------------
In Data Cleaning app, try to apply a custom filter on Company
with "contains" or "not contains" operator.

Issue:
------
A traceback occurs.
For this line: `res = self._obj.execute(query, params)`, we have
`IndexError: list index out of range`.

Cause:
------
There is a mismatch between the `%` in the query and the number of parameters.

The query is constructed by combining subqueries with a template.
These subqueries will be determined using the `mogrify` method and
will be concatenated to the main query template.
The `'` characters are correctly escaped with `\` , but the `%` characters are not.

Solution:
---------
Don't use mogrify, so that we can keep the subquery parameters
and pass them directly to the main query.
By leaving the values wrapped with `%` in the parameters,
there are no more escape problems.

Note:
This is a better fix than manually escaping the `%` if necessary like:
```py
if operator in ('not ilike', 'ilike'):
	value = f'%{value}%'
```

opw-3668433

closes odoo/enterprise#55709

X-original-commit: 4f208f8fcd093047f5162d65ce638e5b9263944f
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>