[IMP] appointment: set default duration at creation
If the appointment type duration is in the context,
set the appointment type duration as the default event
duration when creating a calendar event from the Gantt
view, the Calendar view or from the calendar quick create.

When a user make a selection in the gantt or calendar views,
the start and end of the event is the limit of the selection
no matter if the appointment type is present in the context.

Directly opening the calendar event form view instead of the planning
view for 2 reasons:
- The planning view is unnecessary and kind of strange. Getting the
form view directly is faster and more adequate for creation purpose.
- The gantt view will always give a 59min 59sec type of stop datetime.
Opening the form view instead of the planning view allows to have the
'booking_gantt_create_record' context at creation and thus correctly round
the stop datetime to the nearest minute when coming from the gantt view.

Task-3544915

X-original-commit: 635d9430f43605ba9103e3392878031e757817b3
Part-of: odoo/enterprise#54440