[IMP] {website_}appointment: improve access rules
Purpose
=======
Improve the appointment types and slots access rules as
the internal users had, by default, higher access than
the appointment users.

Specifications
==============
Create an access rule to restrict the internal user access
to only have read access on the:
- published appointment types
- resource based appointment types
- user based appointment types, if there is no staff user or if
the user is in the staff users.
The same logic is applied for the appointment slot.

The appointment users now have read access on every published
appointment types.

Cleaning the fetching of the appointment types by the
"_fetch_and_check_private_appointment_types" method to make use of
the ACLs.

Cleaning the "_prepare_appointments_cards_data" method in
website_appointment to avoid filtering the data twice
(indeed most of the filtering is already done by the
"_fetch_and_check_private_appointment_types" method).

In the appointment_hr module, due to the new rule, the staff_user_bxls
internal user no longer had read access to the test_apt_type.
Setting the 'is_published' to True for the user to have a read access.
This generates a few more queries when calling _get_appointment_slots()
with this appointment type.

Task-3544915

closes odoo/enterprise#54276

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Co-authored-by: amdi-odoo <amdi@odoo.com>
Co-authored-by: jeh-odoo <jeh@odoo.com>