[FIX] timesheet_grid: wrong z-index making dropdown unusable
Im timesheet_grid when starting a timer in the kanban view, the
dropdowns are displayed behind the o_kanban_record, making it impossible
to use.

This is due to a change made on the pinned_header z-index introduced in
commit[1].

This commit applies a z-index-1 on the pinned_header element, making
the dropdowns contained inside this div being able to be displayed above
the kanban cards.

[1]: odoo/enterprise@7b1f18398d0d1f66269d742ef2441ef2bb41f7e6

task-3684460

closes odoo/enterprise#56444

X-original-commit: 100c77b984f3e2721a9a2679473a24f53f22e15e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>