[FIX] appointment: enable translation of xml files on website
Currently some terms we not being translated besides having the term in the .pot file and the translation in the .po files.

Steps to reproduce:
-------------------
* Go to **Setting**
* Add a language (french for example here)
* Enable to translate to your website
* Go to **website**
* Select apointment
* Select any appointment
* Change the language of the website

You can observe that the line *Select a time* is not being translated. When selecting **Translate**, you can observe that the line is not highlight in yellow or green.

Why the fix:
------------
When the route `/website/translations` is called, the modules loaded do not include appointment.

opw-3671043

closes odoo/enterprise#54377

Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>