[FIX] sale_purchase_inter_company_rules: allow PO creation
Steps to reproduce:
- Create two company San Francisco (SF) and Chicago (CH)
- Make a service product, with "Subcontract Services" enabled on both
  company
- For SF set CH as the vendor.
- For CH set an external vendor
- For both company enable "Inter-Company Transaction" and "Automatic
  Validation"
- Enable subcontract services and set the first company as the vendor
- Make a PO on SF and confirm it
- An SO is created on CH

Issue:
A PO should be created on CH but we have nothing.

Solution:
Remove the function override that would stop this behavior.
This code is part of commit (7cb2335), this function restriction was
needed to fix a flow. "Subcontract Services" wasn't company dependent at the
time of the commit, meaning that if company CH didn't have a supplier we
ran into an issue since we would try to confirm a PO without any vendor.
Now that "Subcontract Servives" is company dependent this particular
flow is now avoidable by the client (making it a mis-configuration
instead of a bug).

opw-3622278

closes odoo/enterprise#54316

Related: odoo/odoo#150626
Signed-off-by: William Henrotin (whe) <whe@odoo.com>