[FIX] account_reports: Journal entries missing in report export
Summary:
When we unfold an account and show journal entries in the general
ledger (with the "Hierarchy and Subtotals" option checked), the
journal entries do not appear on the XLSX report.

Steps to reproduce:
  1. Enable debug mode
  2. Create a new account group (Accounting > Configuration >
	Accounting > Account Groups) with a code prefix from 1 to 9
  3. Open the general ledger (Accounting > Reporting > Audit Reports >
	General Ledger) and check the option "Hierarchy and Subtotals"
  4. Unfold a line until you can see journal entries (e.g., "101000
	Current Assets" on runbot)
  5. Export the general ledger in XLSX, and open it: the journal
	entries do not appear

Cause:
In the local stack, there is a field "unfolded_lines" that is a list
containing the unfolded lines in absolute format (e.g., if account A
is in account group G, the line is noted as "G|A").
However, when checking if an account is unfolded, it checks if the
name of the account is in this field (just the name and not the
group/account), which is not the case if the account is in an account
group. In this case, the children are not displayed because only the
content of the unfolded lines is shown.

Fix:
Modify the check for whether an account is unfolded: check if the
name in absolute format is in the unfolded_lines list.

opw-3649550

closes odoo/enterprise#56133

X-original-commit: 98c7ced8d629a1cfbe757a8850a6ccfb9deff86f
Signed-off-by: Cedric Snauwaert <csn@odoo.com>