[IMP] stock_barcode: go back to main menu
In the Barcode App, when an operation or an inventory adjustment is
opened and the page is refreshed, the breadcrumb is lost and it's no
more possible to go back, so the back button does nothing.

With this commit, in case there is no previs action, the back button
will return to the Barcode App main menu.

task-3636518

Part-of: odoo/enterprise#54413