[FIX] web_studio: 'invisible' property on conditional over 'column_invisible'
Before this commit, if setting a condition on a column in studio, it was on the
'column_invisible'. In result, the column was not invisible because it was conditional
on the lines. So you cannot make the column dissapear on some line and not other in
a list view.

After this commit, the property selected when setting a condition is 'invisible'. Like
that the cell is invisible and not the column.

closes odoo/enterprise#58283

Opw-ticket: 3667210
X-original-commit: 63fc1edf4920379b2cb062125e59b0c9e568a8fc
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>