[FIX] appointment: fix traceback when creating a new staff booking in gantt view
This traceback arises when the user tries to create a new 'Staff Booking'
in a gantt view.

To reproduce this issue:

1) Install 'Appointment'
2) Open 'Appointment/Schedule/Staff Booking'
3) Click on 'New' in gantt view of 'Staff Booking'.
4) Traceback arises.

Error:- "TypeError: 'int' object is not subscriptable"

In 'default_get' method, the 'partner_ids' value is given as a list of ids.

https://github.com/odoo/enterprise/blob/43f09c5ed91f012a4caa12379ef913ab77de8d2c/appointment/models/calendar_event.py#L41-L42

Which leads to a traceback from here.

https://github.com/odoo/odoo/blob/86accecf96f63b5047cd70f873e9f17ef35e5331/addons/web/models/models.py#L946-L955

After applying this commit, It will resolve the issue by assigning a value using
command.

sentry-4625709760

closes odoo/enterprise#54580

X-original-commit: cb347b5d97cbc1bfa03469fc71d7c62a5a7da470
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>