[FIX] web_studio: report editor: blacklist uneditable reports
Before this commit, some reports were available to edition for the reportEditor.
Though these reports really do not play well in the report editor as their xml
contain a lot of code and the model on these reports is wrong, meaning that "docs" does not
refer to the actual data that the reports contain eventually.

After this commit, these reports cannot be edited via studio as they are too specific.
In particular, reports that aim at printing labels (stock module and related) are banned via
this commit.

opw-3650049

closes odoo/enterprise#54728

X-original-commit: 40cf7291a8f0fc19a2b9efe02ed4cd341fe3129e
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>