[FIX] social_twitter: fix double mention when we reply to a Tweet
Bug
===
Before, we needed to mention the author of the Tweet we are replying to,
in order to make the Tweet looks like a reply on Twitter. It seems not
needed anymore and causes a double mention issue (only on the Odoo side,
because of the API response, Twitter seems to show only one mention).

Task-3686630

closes odoo/enterprise#56422

X-original-commit: a96c92f028db879d3bc9e3b015c64ef790c51cb9
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>