[FIX] web_enterprise: fix color-list in the Dark Mode
This commit fixes an issue about the `colorlist_item` contrast in dark
mode.

With Odoo 17, the dark mode received an update, and the way we handle
our tags changed. We now use two CSS variables:

1. `--background-color` for the background ;

2. `--color` to handle the colors of the tags.

This approach allow us to have full control on the visual result of our
badges depending on the color scheme used by the user, by using the same
color but with different adjustments.

While this approach was applied to our badges, it was not applied to the
`colorlist_item_color` class, resulting in a readability issue in dark mode.

To fix this issue, we simply add a `--color` CSS variable inside the
color adjustments made to `colorlist_item`, with the same
parameters as the badges one, to ensure that the text/icon inside a
`colorlist` element is always readable.

task-3524295

closes odoo/enterprise#54600

X-original-commit: 86f2594fa2801d8f3c458ed713c8d0e550aeb3c1
Related: odoo/odoo#149911
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Prakash Prajapati (ppr) <ppr@odoo.com>