[IMP] stock_barcode*: main menu button
Since all the Barcode App views are fullscreen even after a reload, that
means after a reloading while on a kanban view, there is no breadcrumb
and no way to go back.
This commit adds a button to go to the Barcode App's main menu on each
of these kanban views.
It's not perfect since this button is folded on small screen and will be
displayed regardless if there is a breadcrumb or not, but it's better
than nothing.

task-3636518

closes odoo/enterprise#54413

Signed-off-by: William Henrotin (whe) <whe@odoo.com>