[FIX] l10n_mx_edi*: Don't fetch twice certificate + fix empty root_company
The certificate was fetched first when checking the config of the company,
then when finding the certificate to use when processing the cfdi.

Also, when we try to send a CFDI when no certificate are configurated, a traceback is raised.
This is because the root_company is the one having some certificate. If no certificate, the
root_company is empty and then, an ensure_one raises an error.

sentry-4719634588,4729692933,4729689997

closes odoo/enterprise#54544

Signed-off-by: Josse Colpaert <jco@odoo.com>