[FIX] l10n_mx_edi: Remove reference to account_edi
[IMP] l10n_mx_edi: Get the Mexican timezone from the same method.
The timezone is hardcoded at multiple places.
It's less error prone to get the timezone everywhere from the same method.

[FIX] l10n_mx_edi: Don't fetch twice certificate + fix empty root_company

The certificate was fetched first when checking the config of the company,
then when finding the certificate to use when processing the cfdi.

Also, when we try to send a CFDI when no certificate are configurated, a traceback is raised.
This is because the root_company is the one having some certificate. If no certificate, the
root_company is empty and then, an ensure_one raises an error.

sentry-4719634588,4729692933,4729689997

closes odoo/enterprise#55390

Forward-port-of: #54544
X-original-commit: e343ada637a55588780b0e97c3af79d98f73cdb5
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>