[FIX] account_reports: don't use old 'print_mode' context key
This context key isn't used anymore. It has been replaced by options['export_mode']. The former calls with the context key corresponded to the generation of pdf exports, and are all converted to 'print', except the one in integer rounding management, for which we rather wish to use the 'file' value. Indeed, when integer rounding is enabled, the pdf must keep the same content as what was shown in the UI, while a file export will always want the legal standard (hence, rounding) to be applied.

closes odoo/enterprise#54561

Signed-off-by: John Laterre (jol) <jol@odoo.com>