[IMP] currency_rate_live: Add support for Bank of Italy currency rates
This commit adds support for the Bank of Italy currency rates. The Bank of Italy publishes the currency rates on its APIs, of which the most recent one is available at https://tassidicambio.bancaditalia.it/terzevalute-wf-web/rest/v1.0/latestRates. Documentation can be found at https://tassidicambio.bancaditalia.it/terzevalute-wf-ui-web/assets/files/Istruzioni_tecnico-operative.pdf.
This is necessary for Italian law compliance, as the Bank of Italy is the only institution allowed to publish currency rates in Italy.

closes odoo/enterprise#71520

X-original-commit: b7e773c0207c19e49ededa43e9f2cf89706a15d8
Signed-off-by: Laurent Smet (las) <las@odoo.com>