[FIX] stock_barcode_picking_batch: scan lots with multiple pickings
How to reproduce:
- Create a product tracked by lots and set a barcode for this product;
- Create a receipt for this product, with at least 2 quantities and
  confirm it;
- Create a second similar receipt (you can duplicate the first one) and
  confirm it;
- Add these two receipt in a new batch and confirm the batch;
- Go in the Barcode App and open the batch, you should have two lines,
  one for each picking;
- Scan the product -> the first line should be selected;
- Scan a lot -> the lot is set on the first line and its qty is
  incremented;
- Scan another lot -> the lot is set on the second line and its qty is
  incremented. That's the issue.

Expected behavior: When another lot is scanned after the first one,
since we are picking quantity for the first picking, the line for the
second picking shouldn't be selected until the first one is processed.
To fix that, the search for a line following the scan will avoid to take
a tracked line if its `picking_id` is not the same than the selected
line and if the selected line is not completed.

task-3688770

X-original-commit: 1c14fb60433987f7dbf7a6f594394e42b3434f1f
Part-of: odoo/enterprise#65182
Related: odoo/odoo#173521
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>