[FIX] web_studio: action context should be transmitted when editing
Before this commit, the context of the action was taken into account
when loading the view the first time one entered studio.

In studio, if one edited the view however, that context was not tramitted, leading
to some confusion when some field's visibility depended on the context.

After this commit, the action's context is passed through to edit_view, which should
be the same as the one that the original get_view was done with.

opw-3687846

closes odoo/enterprise#54623

X-original-commit: 2f4b54f0835687aa7b3f9b7baf1d9df14b674787
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>