[IMP] account_reports: Delayed formatting of values
When displaying a report, all the values need to be formatted in order to display the right number of decimal places, currency symbols, ...

So far, this was done line by line, before prefix groups were possibly applied. On very big reports needing to be trimmed by a prefix group setup, this was a bit stupid, since we formatted all the values, then decided whether to show them or not (then replacing them with prefix lines). In extreme cases (a few thousands of lines), this could even end up with the report taking entire seconds to just format values uselessly.

This commit gets rid of this behavior, and instead does the formatting after generating the lines, when we're sure they need to be displayed.

Task 3059062

Part-of: odoo/enterprise#54594