[IMP] {account,l10n_fr,l10n_nl}_reports: generalize integer rounding to dynamic lines
The integer rounding feature was only available on reports computed from expressions so far, not on custom reports relying on dynamic lines. This was due to the fact this feature had to be introduced in stable, and further refactoring was needed in order to make it work for real.

Now that the formatting of the values has been delayed for all lines (static and dynamic) to the end of the _get_lines, we can add support for integer rounding after the generation of dynamic lines. Since the feature is now available on every report, we also make the integer rounding feature a "real" option, with a field controlling it on the report, and a proper _init_options function.

Task 3059062

closes odoo/enterprise#54594

Related: odoo/odoo#149892
Signed-off-by: John Laterre (jol) <jol@odoo.com>