[FIX] knowledge: fix the clipboard text/plain blob
This commit modifies the text/plain blob used for the template behavior (now
called the Clipboard) because it was using the innerHTML instead of text.
This enabled the user to copy a string containing HTML tags instead of just
text that was pasted inside of the small chatter.

When the user click on the copy button we write inside of the user's clipboard
a blob that contains a ClipboardItem. This Object enables us to choose which
data should be pasted depending on a MIME Type. Here the `text/plain` and
`text/html` were storing the same HTML.
The issue is that when pasting inside an input that uses the `text/plain` type
the user outputs the HTML as litteral string tags and all. Which is not ideal
for the user as it would prefer to have the text content instead of the HTML.

Now this blob will contain the innerText of the template reverting it to its
default behavior.

task-3691185

closes odoo/enterprise#55132

X-original-commit: 3d958f88d4502ea26edee581d91cedccfcdb85f6
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>