[FIX] sale_renting, sale_subscription: check product compatibility
Currently, it is possible to create a product which can be rentable and recurring. This fix intends to replicate the behavior of previous versions where it was impossible to save a product with both boxes enabled.

Steps to reproduce:
-------------------
* Go to **Rental** App
* Select Products
* Create a new product
* Enable Reccuring and Can be Rented
* Save

Why the fix:
------------
https://github.com/odoo/enterprise/commit/2579b5896a4c5e949d30a0a51f18437adc2b5db2 removed the dependence of sale_temporal for sale_subcription. After that, there was no point in keeping the sale_temporal dependence in sale_renting. Thus, the dependence for sale_temporal has also been removed here: https://github.com/odoo/enterprise/commit/6c86a278dbc752009c0985702cb2927c26f5f636. The module sale_temporal was responsible for checking incompatible products types.

Currently, both sale_renting and sale_subscription have the module sale in common. This commit is part of a 2-part fix and targets Odoo Enterprise. It re-introduces the constraints on 'rent_ok' and 'reccuring_invoice'.

Code re-introduced:
* https://github.com/odoo/enterprise/blob/da031dd50501af0b6c02fab804a9af33363d397c/sale_subscription/models/product.py#L12-L19
* https://github.com/odoo/enterprise/blob/da031dd50501af0b6c02fab804a9af33363d397c/sale_renting/models/product_template.py#L19-L26

opw-3678715

closes odoo/enterprise#54940

Related: odoo/odoo#150688
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>