[FIX] hr_payroll_expense: add payroll group for expense sheet field
Issue:
------
A user who has access rights for payroll but not for expenses
will not be able to calculate the payslip.

Cause:
------
An additional security check  [^1] has been added to the `_flush_search` method
to check which fields have groups.
Unfortunately, payroll groups are not mentioned in the `groups`
attribute of the `refund_in_payslip` field.

Solution:
---------
Add the group in the `groups` attribute.

opw-3691235

[^1]: https://github.com/odoo/odoo/blob/4b096631ad23298c12bbd12593e9358c3e9f404a/odoo/models.py#L5330-L5332

closes odoo/enterprise#54619

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>