[FIX] base_import: allow importing readonly fields
Steps to reproduce
------------------
* attempt to import an asset (CSV or XLSX file) with a value for a
  `journal_id` related field.

That value will not be automatically matched with an Odoo Field. If you
try to set it by manually, that field won't appear in the list of
available fields.

Cause
-----
In 16, even though a field was read-only, we would still allow it to be
imported depending on the 'states' field attribute. But since the
removal of that attribute, a read-only field cannot be set trough
import.

Fix
---
This commit makes the assets' `journal_id` field not read-only. The
usual behavior (read-only if not in draft) remains unchanged because the
views were already adapted accordingly.

opw-3635484

closes odoo/enterprise#52872

Signed-off-by: William André (wan) <wan@odoo.com>