[IMP] {l10n_au_}hr_payroll_account: add PAYGW lines to tax report
The Australian government allows companies to report their payroll
activities through a kind of tax report. To fill the tax report lines
representing these activities, we need to update the modules in order to
add tax tags when creating entries from the payslips.

We thought that payroll generated entries could happen more than in one
localisation, so we applied the principle to hr_payroll_account.
Following this idea, we now have tags that can be set directly on the
hr salary rules or hr payslip lines. These tags will be applied on the
corresponding lines when creating the journal entry from the payslip.

For the Australian payroll, we need to be able to change the W2 tag to
W3 by a boolean on the contract, as the default data for salary
structures will represent all the use cases possible (we don't want
users to duplicate structures just to change the tag).

task-2719708

closes odoo/enterprise#48878

Related: odoo/odoo#138553
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>