[FIX] industry_fsm_sale : prevent increase product quantity when editing
Steps to reproduce:
	1- Install Field Service module
	2- Create new task and click on products smart button
	3- Hover over a product and click on the dropdown menu
	4- Click on edit in the dropdown menu
	5- Get back to the products page and check the quantity for the product you edited

Current behavior before PR:
When the user clicks on edit in the dropdown menu of any product the quantity gets increased by 1. This is happening because of the global click event so when the user clicks anywhere inside the kanban box the quantity gets updated.

Desired behavior after PR is merged:
This behavior has been adjusted by checking the target where the user click if it is inside the dropdown menu it will not update the product's quantity.

opw-3689864

closes odoo/enterprise#56951

X-original-commit: 131876052fdc191317547c84079972edad375bb3
Related: odoo/odoo#154643
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>