[IMP] helpdesk, *: add missing editor message to know is shared area
* website_helpdesk

When the `oe_structure_<id>` ID is added inside a shared view (rendered
on multiple records pages), it needs to inform the user that the
elements drag & drop here will be replicated on all other pages.
Otherwise, the user has no clue what's going on, he just sees some areas
on a record page, but when saved it is applied on all record pages.

Some occurences were missing that hint.

opw-3671051

Part-of: odoo/enterprise#55644