[IMP] l10n_au_hr_payroll: add Termination payment to the last payslip
The termination payments are added as a separate
salary structure. That only allows termination payment
to be made through a new payslip after the final payslip.

This commit allows the rules from the temination structure
to be added directly to the last payslip along with the default
structure.

task # 3639270

Part-of: odoo/enterprise#57884