[FIX] l10n_pl_reports: upgrade to 17.0 fails because of xml_ids
Because of the renaming of the some lines in https://github.com/odoo/enterprise/commit/272856993b21d9e7edfb573e33a368f80ba0c9c8,
there are conflicts between old and new lines.
For l10n_pl_small_bs_assets_a_3, we had a shortcut for the expression.
We need to nullify it, if not the field stays filled on the model.

For l10n_pl_small_bs_assets_a_4, the problem is bigger.
We had an aggregation before and we go to a line with groupby.
The problem is that we add the groupby before the suppression of
expressions.
We can't nullify a field to correct it, so we just change the
ids back.

Linked to runbot error 55641, 55642, 55643, 55644

closes odoo/enterprise#55159

X-original-commit: aa9369da02944b898e3c4489bd11d165baf4f88e
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>